homework_1/pyproject.toml
[project]
name = "ai-agents-homework-1"
version = "0.1.0"
description = "Homework 1: LLM API with Calculator Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py", "test_main.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"